Experimental refactor. Project integration work.
