--- cmake/LibHdb++Config.h.in
@@ -1,9 +1,0 @@
-#ifndef LIBHDPPP_CONFIG_H
-#define LIBHDPPP_CONFIG_H
-
-#define VERSION_MAJOR "${LIBHDBPP_VERSION_MAJOR}"
-#define VERSION_MINOR "${LIBHDBPP_VERSION_MINOR}"
-#define VERSION_REVISION "${LIBHDBPP_VERSION_REVISION}"
-#define BUILD_TIME "${LIBHDBPP_TIMESTAMP}"
-
-#endif /* LIBHDPPP_CONFIG_H */
