--- cmake/ReleaseVersion.cmake
@@ -1,13 +1,0 @@
-# Project version
-set(LIBHDBPP_VERSION_MAJOR "1")
-set(LIBHDBPP_VERSION_MINOR "0")
-set(LIBHDBPP_VERSION_REVISION "0")
-string(TIMESTAMP LIBHDBPP_TIMESTAMP "%Y-%m-%d %H:%M:%S")
-
-# Version the shared library
-set(LIBRARY_VERSION_MAJOR 6)
-set(LIBRARY_VERSION_MINOR 0)
-set(LIBRARY_VERSION_PATCH 0)
-
-set(LIBRARY_VERSION_STRING 
-    ${LIBRARY_VERSION_MAJOR}.${LIBRARY_VERSION_MINOR}.${LIBRARY_VERSION_PATCH})
